+++ src/main/webapp/wise5/components/discussion/i18n/i18n_en.json
@@ -6,5 +6,6 @@
   "discussion.componentTypeLabel": "Discussion",
   "discussion.post": "Post",
   "discussion.repliedToADiscussionYouWereIn": "{{userNames}} replied to a discussion you were in!",
-  "discussion.reply": "Reply"
+  "discussion.reply": "Reply",
+  "discussion.shareDiscussionPostWithClass": "Share with class..."
 }
